fully functional and Valgrind-happy communication function
The `uca_phantom_communicate` function is now capable of requesting
unit variables in a, as of yet, memory-safe way. No leaks detected,
other than the possible leaks generate by GObject/GLib itself.

TODO: The highest item in the priority list is to polish up the
error handling and implement the UcaPhantomCommunication GObject
class.