Wrap the network functions into a GObject class like a burrito
Implemeted a GObject derived final class UcaPhantomCommunicate
class with all of the relevant network functions. This will help
with network-state tracking and will also help to encapsulate
all the necesarry public functions for the user. i.e. further
resrict a random user from toying with any random function.
Finally, the next step is to integrate the custum Uca Phantom
errors. The goal is to make the error handling as robust as possible.

Note: all commits are memory-leak safe as of yet.