Add custom phantom communicate error domain + add set variable func
Custom errors have been implemented for better error tracing. A small
error-propagation bug was also fixed: when propagating the error to
the caller function, the data belongs to the caller and not the base
function. That is why one should not free the error once it is propagated.

Moreover, a uca_phantom_set_variable function was added. The set/get
unit variable functions are still synchronous.
Future feature to think about: implementing as async version/buffered/
throttled.