add 1Gb network image retrieval functions
Yipee! The code can now succesfully read batches of n images from
over the 1Gb ethernet connection. The camera first has to be
triggered, and then you can retrieve as many images as you want
(so long as its less that the amount of post trigger frames).
A number of other functionalites have been implemented; most
notably, one can now control the internal parameters of the
camera using a struct called `CaptureSettings`. With this, you
can control the framerate, exposure time, etc.

Once the code will be fully integrated to libuca, this struct
wwill probably be useless since these properties will be set
through the `set/get` mechanisms...

The retrieved batches of images are stored in structs that contain
the properties of the recording. These structs are contained in
an `GAsyncQueue`. There are no ring buffers beeing used.

The unpacking of these binary images has not been implemented in
this commit. But this is less urgent than for example retrieving
images overs the 10Gb ethernet, as the images can always be unpacked
afterwards. (as is done right now)