Fixed export of GObject properties for ucad-net and improved variable naming
The image related variables such as sensor_pixel_width used to designate the
sensor resolution. Conform to libuca, this is now the physical size of the pixel
in m. The same can be said for the related variables.

Moreover, the installation of the properties has been improved to work with ucad.
Although I'm not too sure how they started working; it seems as though overriding
the name base class property as a user property made things work...