Support for early images feature of camera
Because of how the camera indexes the frames after a trigger, one must wait a little for the cine buffer to fill up before triggering, so that the first few frames are indeed indexed from 0. Without waiting, the first few start at a positive integer and one will have to go back and fetch those images once all the post frames are done.