Bug fix: blocking data connection from camera to server
I am not 100% certain of my explaination, but it seems like
since the uca-phantom instance was running in the GMainLoop
of ucad.c, all the signals were caught in this loop. In particular,
I was using g_socket_listener_accept instead of using a signal,
and it seems that ucad was catching this signal and blocking
g_socket_listener_accept indefinitely as it could never get the
connection. Regardless, using the incoming signal works now.