Openmp fully functional
Here's a list of what's been introduced:
- New openmp based unpacker: on my system, easily unpack 12bit data at
10Gbytes/sec
- Ringbuf has new functionality: Direct Memory Access with reserve and commit
function. This allows for some nice multithreaded work on the buffer without
compromising speed.
- Fixed confusion regarding resolution properties. ROI is ROI, window is the
output ROI (ok that's confusing... But it doesn't work and won't be used), and
sensor width/height is the resoltion (which IMO is also confusing, but let's
not go break the libuca api dependencies). I determined that the ROI
functionality integrated in the phantom does not seem to work. Still, it is
implemented and can be tested.
- Added a mempool option for situations where you need to absolutely cap the
memory usage of the program. For example you only have 2Gb left on your system,
you could enable the USE_MEMPOOL option, and cap it at 2Gb. This would give the
driver 2Gb of leeway to receive the images at 10Gb rate. Note that this should
hinder the performance a little, but not dramatically. Also ideal for embedded
systems.
- Fixed some leaks here and there, improved error checking.